Set tags on created roles
Until now, we hadn't been setting any tags on the roles we create.  Here
we start setting them from `module.label.tags`, merging in a new
`var.additional_role_tags` to allow for specific tags to be set only on
the role itself.

We also allow a couple additional options to be set on
`aws_iam_role.default`: `path` and `max_session_duration`.